+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.9.1] - 2026-8-18
+
+- Fix sharing scene order
+- Fix PDF export crash
+- Fix MacOS crash on start
+- Added configurable timezone for server admins
+
+
 ## [3.9.0] - 2026-8-16
 
 New features